MCP lint/diagnostic_summary now run Erlang FFI argument-type checks (BT-2858)
The MCP `lint`/`diagnostic_summary` tools' analysis pipeline
(`run_module_analysis`) always analysed with `native_type_registry: None`,
so it never ran FFI argument-type checks or the typed FFI return-type
inference they depend on (ADR 0075) — silently diverging from `beamtalk
lint`/`beamtalk build` for any code calling into Erlang FFI.

Moves the OTP/dependency `.beam` type-spec extraction logic
(`extract_type_specs`) from the `beamtalk-cli` binary's `beam_compiler`
module into the shared `beamtalk_cli` library crate (`native_type_specs`),
mirroring the `dependency_classes`/`build_layout`/`manifest` moves done for
BT-2823/BT-2836, so `beamtalk-mcp` can call the exact same extraction code
`beamtalk lint`/`beamtalk build` use instead of reading a possibly-stale
on-disk cache. `commands::build::extract_type_specs` now delegates to the
lib function; existing `beam_compiler::X` call sites are preserved via a
`pub use` re-export.
