+++ crates/beamtalk-cli/src/lib.rs
@@ -10,9 +10,12 @@
 //! - [`build_layout`] — centralised `_build/` path construction
 //! - [`dependency_classes`] — best-effort, offline dependency class
 //!   resolution shared with `beamtalk-mcp` (BT-2823)
+//! - [`native_type_specs`] — Erlang FFI type-spec extraction shared with
+//!   `beamtalk-mcp` (BT-2858)
 
 pub mod build_layout;
 pub mod dependency_classes;
 pub mod erlc;
 pub mod manifest;
+pub mod native_type_specs;
 pub mod repl_startup;
