+++ docs/development/surface-parity.md
@@ -207,7 +207,7 @@ These CLI subcommands are build/tooling commands that operate offline (no worksp
 | `check` | -- | `textDocument/publishDiagnostics` | Offline syntax/type check; LSP provides diagnostics on save. **BT-2800 (ADR 0100 Rule 3):** both `beamtalk build`/`check` and the LSP apply the package's `beamtalk.toml` `[diagnostics]` per-category severity-override table through the same shared pipeline (`compute_project_diagnostics` → `apply_diagnostics_table`, `beamtalk-core`'s `compilation::diagnostics_policy` module) — a package that sets `dnu = "error"` fails the CLI build *and* shows the same site as an `Error` in the editor, closing what was previously a surface-parity gap (a `beamtalk-lsp → beamtalk-cli` dependency was avoided by moving the table parser + `apply_diagnostics_table` into `beamtalk-core`, which both crates already depend on). The LSP loads each workspace root's table once at startup (`Backend::load_diagnostics_table`, alongside the ADR 0075 type-cache load) — editing `beamtalk.toml` while the server is running requires a restart to pick up the change. **BT-2839 extended this to the REPL**, closing a third instance of the same gap: the REPL's diagnostics entry point (`compute_diagnostics_with_known_vars_and_classes`, called from `beamtalk-compiler-port`'s `compile_expression`/`compile`/`diagnostics` request handlers) now applies the same table too, so `dnu = "error"` fails `beamtalk build`, shows `Error` in the LSP, *and* shows `Error` at the REPL — never a REPL-only soft `Hint`. The compiler-port process caches the table once per session, read from `beamtalk.toml` in its working directory (which the parent BEAM node pins to the project root at REPL startup) — mirroring the LSP's load-once-at-startup semantics; like the LSP, a `beamtalk.toml` edit requires restarting the REPL session to take effect. **Known gap (BT-2859):** unlike `beamtalk lint` (fixed in BT-2851), the LSP's `Backend::load_type_cache` only reads `_build/type_cache/`; it never falls back to live `.beam` extraction. On a workspace opened before any `beamtalk build` has run, its `NativeTypeRegistry` stays empty, so Erlang FFI argument-type diagnostics `beamtalk build`/`beamtalk test` produce are absent from the editor, and an `@expect type` suppressing one shows as stale until the user builds and reloads. |
 | `fmt` | -- | `textDocument/formatting` | Format source files; LSP provides document formatting |
 | `fmt-check` | -- | -- | `surface-specific: CI formatting check` |
-| `lint` | `lint` | -- | Lint checks; MCP exposes for AI-assisted workflows. **BT-2823:** both resolve dependency (`beamtalk.toml`) class metadata into the same `ClassHierarchy` so `Unresolved class` diagnostics match for classes defined only in a git/path dependency. Deliberate divergence: the CLI's `resolve_dep_class_infos` will fetch a missing/stale git dependency over the network (`ensure_deps_resolved`); the MCP tool stays fully offline and only reads whatever dependency checkout already exists under `_build/deps/<name>/` (as left by a prior `beamtalk build`) — an unfetched dependency's classes are silently skipped rather than triggering a fetch. **BT-2851:** the CLI's `beamtalk lint` now populates its Erlang FFI native-type registry via the same `extract_type_specs` extractor `beamtalk build`/`beamtalk test` call (falling back to live `.beam` extraction when `_build/type_cache/` is absent or stale, rather than only reading whatever the cache holds), so `@expect type` on an FFI argument-type mismatch is never stale on one surface and legitimate on the other. Known gap, not fixed here: the MCP `lint` tool's analysis path (`run_module_analysis` in `crates/beamtalk-mcp/src/server.rs`) does not pass a native-type registry at all, so it never runs FFI argument-type checks or the return-type inference they depend on — tracked in BT-2858. |
+| `lint` | `lint` | -- | Lint checks; MCP exposes for AI-assisted workflows. **BT-2823:** both resolve dependency (`beamtalk.toml`) class metadata into the same `ClassHierarchy` so `Unresolved class` diagnostics match for classes defined only in a git/path dependency. Deliberate divergence: the CLI's `resolve_dep_class_infos` will fetch a missing/stale git dependency over the network (`ensure_deps_resolved`); the MCP tool stays fully offline and only reads whatever dependency checkout already exists under `_build/deps/<name>/` (as left by a prior `beamtalk build`) — an unfetched dependency's classes are silently skipped rather than triggering a fetch. **BT-2851:** the CLI's `beamtalk lint` now populates its Erlang FFI native-type registry via the same `extract_type_specs` extractor `beamtalk build`/`beamtalk test` call (falling back to live `.beam` extraction when `_build/type_cache/` is absent or stale, rather than only reading whatever the cache holds), so `@expect type` on an FFI argument-type mismatch is never stale on one surface and legitimate on the other. **BT-2858:** the MCP `lint`/`diagnostic_summary` tools' analysis path (`run_module_analysis` in `crates/beamtalk-mcp/src/server.rs`) now builds and passes the same registry (`build_native_type_registry`, which calls `beamtalk_cli::native_type_specs::extract_project_type_specs` — the extraction logic itself moved from the CLI binary's `beam_compiler` module into the shared `beamtalk_cli` library crate so both surfaces call the exact same code, not just an equivalent copy), closing the last MCP/CLI divergence on FFI argument-type checks and return-type inference. |
 | `test-script` | -- | -- | `surface-specific: btscript expression tests (CI)` |
 | `test-docs` | -- | -- | `surface-specific: doctest runner (CI)` |
 | `doctor` | -- | -- | `surface-specific: environment health check` |
